Move EQPreset.cs to Clickwheel.DataTypes
